tetragon: Move ratelimit maps and defs to separate file
This way it's easier to include it from other objects.
